==> 7_CombineSentimentAndImageLabels.ipynb <==
import pandas as pd

df_images = pd.read_csv("data/silver_20250322_Airline_Reviews_ImageCLIPembeddings.csv")
df_images

df_sentiment = pd.read_csv('data/silver_20250322_Airline_Reviews_Sentiment.csv')
df_sentiment

# Remove this when 2_DataPreprocessing step is added
df_sentiment.rename(columns={"Unnamed: 0": "RowId"}, inplace=True)
df_sentiment

#### Merge image sentiment labels

image_columns_needed = ['RowId','img_resnet50_score','img_resnet50_label','img_blip_caption',\
    'img_clip_embedding','img_caption_clip_embedding','img_resnet50label_clip_embedding',\
    'img_vs_caption_similarity','img_caption_vs_resnetlabel_similarity']

df_merged = pd.merge(df_sentiment, df_images[image_columns_needed], on='RowId', how='left')
df_merged

#### Persist to CSV file

df_merged\
    .to_csv("data/silver_20250322_Airline_Reviews_SentimentImageLabelsMerged.csv", index=False)
